Create `MouseEventController` which extends `BaseController` with `MouseEventDispatcher` methods.
