+++ package.json
@@ -11,7 +11,8 @@
     "jquery": "./code/js/lib/jquery-1.11.1.min.js",
     "ko": "./node_modules/knockout/build/output/knockout-latest.js",
     "lodash": "./node_modules/lodash/index.js",
-    "BaseController": "./code/js/modules/BaseController.js"
+    "BaseController": "./code/js/modules/BaseController.js",
+    "MouseEventController": "./code/js/modules/MouseEventController.js"
   },
   "browserify-shim": {
     "jquery": "jQuery"
